pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.7.1",
    "py-cord>=2.7.1",
    "python-dotenv>=1.2.2",
]
